Extract cache handling into its own file
