+++ app/src/main/java/com/katiearose/sobriety/Main.kt
@@ -16,12 +16,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import java.io.FileNotFoundException
-import java.io.InputStream
-import java.io.ObjectInputStream
-import java.io.ObjectOutputStream
 import java.time.Instant
-import java.util.zip.DeflaterOutputStream
-import java.util.zip.InflaterInputStream
 
 
 class Main : AppCompatActivity() {
@@ -51,7 +46,7 @@ class Main : AppCompatActivity() {
     private lateinit var prompt: TextView
 
     private lateinit var adapterAddictions: AddictionCardAdapter
-
+    private lateinit var cacheHandler: CacheHandler
     private val createCardRequestCode = 1
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -61,10 +56,10 @@ class Main : AppCompatActivity() {
         addCardButton = findViewById(R.id.addCardButton)
         addCardButton.setOnClickListener { newCardDialog() }
         prompt = findViewById(R.id.prompt)
-
+        cacheHandler = CacheHandler(this)
         try {
             this.openFileInput("Sobriety.cache").use {
-                readCache(it)
+                addictions.addAll(cacheHandler.readCache(it))
             }
         } catch (e: FileNotFoundException) {
         }
@@ -77,7 +72,6 @@ class Main : AppCompatActivity() {
         val layoutManager = LinearLayoutManager(this)
         recyclerAddictions.layoutManager = layoutManager
         recyclerAddictions.adapter = adapterAddictions
-
         //main handler to refresh all cards in sync
         val mainHandler = Handler(Looper.getMainLooper())
         mainHandler.postDelayed(object : Runnable {
@@ -87,6 +81,7 @@ class Main : AppCompatActivity() {
                 if(!deleting){
                     adapterAddictions.notifyDataSetChanged()
                 }else{
+                    cacheHandler.writeCache()
                     deleting = false
                 }
                 mainHandler.postDelayed(this, 1000L)
@@ -129,51 +124,10 @@ class Main : AppCompatActivity() {
         }
     }
 
-    private fun readCache(input: InputStream) {
-        val cache = input.readBytes()
-        try {
-            InflaterInputStream(cache.inputStream()).use { iis ->
-                ObjectInputStream(iis).use {
-                    addictions.addAll(it.readObject() as ArrayList<Addiction>)
-                }
-            }
-        } catch (e: ClassCastException) {
-            readLegacyCache(cache.inputStream())
-        }
-    }
-
-    private fun readLegacyCache(input: InputStream) {
-        try {
-            val a = HashMap<String, Pair<Instant, CircularBuffer<Long>>>()
-            InflaterInputStream(input).use { iis ->
-                ObjectInputStream(iis).use {
-                    for (i in it.readObject() as HashMap<String, Pair<Instant, CircularBuffer<Long>>>) {
-                        a[i.key] = i.value
-                    }
-                }
-            }
-            for (ad in a) {
-                val addiction = Addiction(ad.key, ad.value.first)
-            }
-        } catch (e: Exception) {
-            //Do nothing, i.e. if the cache is older than the previous version just ignore it, supporting every previous version would take more code than it's worth.
-        }
-    }
-
-    private fun writeCache() {
-        this.openFileOutput("Sobriety.cache", MODE_PRIVATE).use { fos ->
-            DeflaterOutputStream(fos, true).use { dos ->
-                ObjectOutputStream(dos).use {
-                    it.writeObject(addictions)
-                }
-            }
-        }
-    }
-
-    override fun onSaveInstanceState(outState: Bundle) {
-        super.onSaveInstanceState(outState)
-        writeCache()
-    }
+//    override fun onSaveInstanceState(outState: Bundle) {
+//        super.onSaveInstanceState(outState)
+//        writeCache()
+//    }
 
     /**
      * This gets called once the Create Activity is closed (Necessary to hide the prompt in case
@@ -192,6 +146,7 @@ class Main : AppCompatActivity() {
                 val instant = data.extras?.get("instant") as Instant
                 val addiction = Addiction(name, instant)
                 addictions.add(addiction)
+                cacheHandler.writeCache()
                 adapterAddictions.notifyDataSetChanged()
             }
         }
